Adds "state.[h,c]" to help decomposing injector operations.
single_injection_run is getting complicated, and hard to extend. A
necessary precondition for splitting it into smaller pieces is packaging
the state so that we don't have to pass around 10 parameters everywhere.

This is a first step in that direction. Now the OPEN and "is_dir" checks
can be broken out into their own helper function, for example.