Investigation into getting backtracer to work inside injector.
Unfortunately, we've discovered that using PTRACE_SYSCALL makes all signals
appear to the tracer as SIGTRAPS, so it can't easily distinguish segfaults, etc.
For some reason it also results in us having a bizarre %rip when we try to do
the backtrace of the child process.

If we use PTRACE_CONT instead, we can successfully backtrace on our segfault_target.c,
but then we lose the ability to intercept system calls (not ok!)

It's not clear how to make PTRACE_SYSCALL and our backtracer work nicely together...